==> 2-Network Explorer.Rmd <==
---
title: "Network Explorer"
author: "Andrés Castro Araújo"
date: "3/14/2019"
output: 
  html_document:
    code_folding: hide
    theme: paper
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE, 
                      fig.align = "center")
```

This notebook goes over the steps necessary to create [**THIS**](https://acastroaraujo.shinyapps.io/FDI_network_map/) Shiny app.

```{r}
library(tidyverse)
library(countrycode)
library(geonames)

source_info <- read_csv("source_info.csv")
target_info <- read_csv("target_info.csv")
gdp <- readRDS("wdi_data/wdi_data.RDS") %>% 
  select(country, continent, year, NY.GDP.MKTP.CD) %>%
  rename(gdp = NY.GDP.MKTP.CD) %>% 
  ungroup()

target_info <- target_info %>%
  left_join(rename(gdp, target = country)) %>% 
  rename(target_gdp = gdp, target_continent = continent) %>% 
  drop_na() %>% 
  left_join(rename(gdp, source = country)) %>% 
  rename(source_gdp = gdp, source_continent = continent) %>% 
  drop_na()

source_info <- source_info %>%
  left_join(rename(gdp, target = country)) %>% 
  rename(target_gdp = gdp, target_continent = continent) %>% 
  drop_na() %>% 
  left_join(rename(gdp, source = country)) %>% 
  rename(source_gdp = gdp, source_continent = continent) %>% 
  drop_na()
```

**Get "sizes" for each country and year**

```{r}
source_info_sizes <- source_info %>% 
  mutate(ff = flow * 1e6 / target_gdp) %>% 
  group_by(target, year) %>% 
  summarise(size = sum(ff)) %>% 
  arrange(-size)

target_info_sizes <- target_info %>% 
  mutate(ff = flow * 1e6 / target_gdp) %>% 
  group_by(target, year) %>% 
  summarise(size = sum(ff)) %>% 
  arrange(-size)
```

**Get world map**

```{r}
library(maps)
library(mapproj)
world_data <- map_data("world") %>% 
  filter(region != "Antartica") %>% 
  mutate(iso = maps::iso.alpha(region, n = 3))

  

theme_map <- theme_void(base_family = "Palatino") + 
  theme(plot.background = element_rect(fill = "antiquewhite"))

mapcoords <- coord_fixed(xlim = c(-170, 180), ylim = c(-55, 80))

country_shapes <- geom_polygon(aes(x = long, y = lat, group = group),
                               data = world_data,
                               fill = "steelblue1", color = "grey",
                               size = 0.05) 

g <- ggplot(NULL) + country_shapes + theme_map + mapcoords
g
```

**Get country longitudes and latitudes**

```{r}
source("google_key_api.R")
library(ggmap)
register_google(key)

COL <- geocode("Colombia")
COL

g + geom_point(data = COL, aes(x = lon, y = lat), color = "red")
```

```{r, eval = FALSE}
library(igraph)
Gt <- graph_from_data_frame(target_info)
nodes <- tibble(name = V(Gt)$name)

nodes <- nodes %>% 
  mutate(iso = countrycode::countrycode(
    name, 
    origin = "country.name", 
    destination = "iso3c")
    )

nodes <- nodes %>% 
  mutate(coords = purrr::map(paste(name, "country"), geocode))

nodes <- nodes %>% 
  unnest(coords)

write_delim(nodes, "nodes.txt")
```

```{r}
node_info <- read_delim("nodes.txt", delim = " ")
```

```{r}
g + geom_point(data = node_info, aes(x = lon, y = lat), size = 0.5)
```

**Note**. *Some of these might be wrong (Google can make mistakes)!*

```{r}
library(ggraph)
library(igraph)
library(tidygraph)

get_iso <- function(x) {
  countrycode::countrycode(
    x, origin = "country.name", 
    destination = "iso3c")
}

make_flows <- function(country = "Eritrea", y = 2012, info = "target", direction = "inflow") {
  theme_map <- theme_void(base_family = "Palatino") + 
    theme(plot.background = element_rect(fill = "white", colour = "white"),
          plot.title = element_text(hjust = 0.5))
  
  if (direction == "inflow") {
    type <- quote(target)
    other <- quote(source)
  }
  
  if (direction == "outflow") {
    type <- quote(source)
    other <- quote(target)
  }

  if (info == "target") {
    df <- target_info %>% 
      filter(!!type == country, year == y, flow != 0) %>% 
      mutate(iso = get_iso(!!other),
             edge_weight = abs(flow) * 1e5 / 
             ifelse(type == "inflow", target_gdp, source_gdp)
             ) 
  }
  if (info == "source") {
    df <- source_info %>% 
      filter(!!type == country, year == y, flow != 0) %>% 
      mutate(iso = get_iso(!!other),
             edge_weight = abs(flow) * 1e5 / 
             ifelse(type == "inflow", target_gdp, source_gdp)
             ) 
  }
  
  if (nrow(df) == 0) {
    return("No data available for this selection.")
  }
  
  Gt <- graph_from_data_frame(df) %>% 
    as_tbl_graph() %>% 
    activate(nodes) %>% 
    left_join(node_info) %>% 
    activate(edges) %>% 
    mutate(active = case_when(
      flow < 0 ~ "negative",
      flow > 0 ~ "positive"
    ))
  
  lay <- create_layout(
    graph = Gt, 
    layout = "manual",
    node.positions = rename(
      as_data_frame(Gt, what = "vertices"), 
      x = lon, y = lat)
    )
  
  world_df <- world_data %>% 
    left_join(df) %>% 
    mutate(active = case_when(
      flow < 0 ~ "negative",
      flow > 0 ~ "positive",
      iso == get_iso(country) ~ "focus"
    ))
  
  country_shapes <- geom_polygon(
    aes(x = long, y = lat, group = group, 
        fill = active, alpha = edge_weight),
    data = world_df,
    color = "grey",
    size = 0.05,
    show.legend = FALSE)
  
  g <- ggraph(lay) + 
    country_shapes +
    geom_edge_arc(aes(color = active, alpha = edge_weight), 
                  curvature = 1/3, show.legend = FALSE,
                  arrow = arrow(length = unit(0.1, "cm")),
                  end_cap = circle(0.1, "cm"),
                  start_cap = circle(0, "cm"),
                  width = 0.8) +
    theme_map + mapcoords
  
  if (any(df$flow < 0, na.rm = TRUE) & any(df$flow > 0)) {
    g <- g + 
      scale_edge_color_manual(values = c("red", "green3")) +
      scale_fill_manual(values = c("black", "red", "green3"))
  } else if (any(df$flow > 0)) {
    g <- g + 
      scale_edge_color_manual(values = c("green3")) +
      scale_fill_manual(values = c("black", "green3"))
  } else if (any(df$flow < 0)) {
    g <- g +
      scale_edge_color_manual(values = c("red")) +
      scale_fill_manual(values = c("black", "red"))
  }
  
  if (direction == "inflow") {
    g <- g + 
      labs(title = paste0("Foreign Direct Investment to ", country, ", ", y))
  }
  
  if (direction == "outflow") {
    g <- g + 
      labs(title = paste0("Foreign Direct Investment from ", country, ", ", y))
    }  
  return(g)
}
```

```{r}
make_table <- function(country = "Cayman Islands", y = 2005, info = "source", direction = "inflow") {
  if (direction == "inflow") {
    type <- quote(target)
    other <- quote(source)
  }
  
  if (direction == "outflow") {
    type <- quote(source)
    other <- quote(target)
  }
  
  if (info == "target") {
    output <- target_info %>% 
      filter(!!type == country, year == y, flow != 0) %>% 
      select(!!other, !!type, flow) %>% 
      arrange(-abs(flow))
  }
  
  if (info == "source") {
    output <- source_info %>% 
      filter(!!type == country, year == y) %>% 
      select(!!other, !!type, flow) %>% 
      arrange(-abs(flow))
  }
  return(output)
}


make_table()
```

```{r}
make_flows("Japan", 2012, info = "source", direction = "inflow")
make_table("Japan", 2012, info = "source", direction = "inflow")
```


```{r}
make_flows(country = "Congo - Brazzaville", y = 2010, info = "target", direction = "inflow")
make_table("China", 2010, "target", "inflow")



make_flows("Eritrea", 2012, "target", "inflow")

```
